Move 2pc implementation into separate module
